Moved cryptography patch to Tribler core
