+++ Tribler/Main/hacks.py
@@ -42,33 +42,6 @@
 
 import sys
 
-
-def patch_crypto_be_discovery():
-
-    """
-    Monkey patches cryptography's backend detection.
-    Objective: support pyinstaller freezing.
-    """
-
-    # Just do the monkeypatching if running on a windows installer version.
-    if (sys.platform == 'win32' and sys.argv[0].lower().endswith("tribler.exe")) \
-            or (sys.platform == 'darwin' and hasattr(sys, 'frozen')):
-        from cryptography.hazmat import backends
-
-        try:
-            from cryptography.hazmat.backends.commoncrypto.backend import backend as be_cc
-        except ImportError:
-            be_cc = None
-
-        try:
-            from cryptography.hazmat.backends.openssl.backend import backend as be_ossl
-        except ImportError:
-            be_ossl = None
-
-        backends._available_backends_list = [
-            be for be in (be_cc, be_ossl) if be is not None
-        ]
-
 # Python2 sys.argv uses the non-unicode windows API, so non-ascii chars will be mangled.
 # TODO(emilon): This should be removed once we move to Python3
 # Based on code from:
